File reorganization/renaming
The particle class was, to put it nicely, unmanagable. The thing had grown to 800+ lines. Trying to find a particular function in that mess was a task on its own. To fix this, I decided to move all of the particle 'helper' methods into their own files.

I defined a new namespace, the 'Particle_Helper' namespace. This namespace includes basically all functions that accept particles as argumnets/help perform the particle iterations. This namespace can be seen in Particle_Helper.h

You may notice that there is no Particle_Helper.c file. That is because I decided to split the function definitions over several files, organizing the functions by functional usage.

I made a source file for the Neighbor functions (Are_Neighbors, Find_Neighbors, Find_Neighbors_Box) called 'Particle_Neighbors.c'.

I made a source file for the Particle update methods (Update_P, Update_x) called 'Particle_Update.c'

I made a source file for the Particle_Damage methods (Removed_Damaged_Particle) called 'Particle_Damage.c'

Finally, I made a source file for the soon to be created 'Contact force' methods called 'Particle_Contact.c'.

The methods in the Particle_Helper namespace have not been changed. Rather, the namespace and new files  just gives me a way of organizing the method definitions + hiding the methods from functions that shouldn't be using them.

I also renamed all .c files to .cc. This was done so that c++ keywords are highlighted (this should have been done a LONG time ago).

The next step is to define the particle contact methods. For these, damage will be disabled.
