+++ Particle_Contact.cc
@@ -1,0 +1,9 @@
+#if !defined(PARTICLE_CONTACT)
+#define PARTICLE_CONTACT
+
+#include "Particle_Helpers.h"
+
+void Particle_Helpers::Contact(Particle * Body_A, const unsigned int Num_Particles_A, Particle * Body_B, const unsigned int Num_Particles_B) {
+} // void Particle_Helpers::Contact(Particle * Body_A, const unsigned int Num_Particles_A, Particle * Body_B, const unsigned int Num_Particles_B) {
+
+#endif
