+++ SPH_Diagnostics.cc
